feat(ci): golangci-lint(v2) 導入 + errorlint/gosec 指摘を解消 (#1796)
* feat(ci): golangci-lint(v2) 導入 + errorlint/gosec 指摘を解消

品質向上の一環。default の staticcheck/errcheck/govet/unused/ineffassign に加え
errorlint / bodyclose / gosec / misspell / unconvert を有効化し、既存指摘を解消した。

- backend/.golangci.yml 新規。gosec の誤検知は理由付きで除外(自作 lint ツールの
  WalkDir / テストの dir 権限・ダミー認証情報 / サンドボックスの意図的 exec=G204)
- ci-backend-go.yml の単体 staticcheck step を golangci-lint-action(v2)に置換
  (staticcheck は golangci に内包)
- errorlint 解消:
  - handler の switch on error / テストの != 比較 → errors.Is
  - fmt.Errorf("%w: %v", ...) → "%w: %w"(原因 error もラップ)
  - exec.ExitError の type assertion → errors.As
- gosec G115 解消: AI チャット SSE の sessionId を int64→uint64 にし変換を除去
  (負値は bind 時に弾かれる)

go test ./... green / golangci-lint 指摘 0。


* fix(ci): gosec G204 除外を execute_code_usecase.go に限定

CodeRabbit 指摘対応。G204 をグローバル除外すると将来別箇所の危険な exec を
見逃すため、サンドボックス(execute_code_usecase.go)に限定した exclusions.rules に変更。
他箇所の exec.Command は引き続き gosec G204 で検出される。


---------

+++ backend/internal/handler/ai_chat_sse_handler.go
@@ -28,7 +28,7 @@ func NewAiChatSseHandler(sendStream *usecase.SendAiMessageStreamUseCase) *AiChat
 }
 
 type sseRequestBody struct {
-	SessionID   int64                  `json:"sessionId"`
+	SessionID   uint64                 `json:"sessionId"`
 	Content     string                 `json:"content"`
 	Scene       string                 `json:"scene"`
 	SessionType string                 `json:"sessionType"`
@@ -103,7 +103,7 @@ func (h *AiChatSseHandler) Handle(c *gin.Context) {
 
 	stream, err := h.sendStream.Execute(ctx, usecase.SendAiMessageInput{
 		UserID:      uid,
-		SessionID:   uint64(body.SessionID),
+		SessionID:   body.SessionID,
 		Content:     body.Content,
 		Scene:       body.Scene,
 		SessionType: body.SessionType,
